+++ react/admin/OrganizationsList/OrganizationsListSearch/OrganizationsListSearch.tsx
@@ -1,8 +1,10 @@
 import React from 'react'
 import { Search, csx, useSearchState } from '@vtex/admin-ui'
 import type { TagProps } from '@vtex/admin-ui'
+import { useIntl } from 'react-intl'
 
 import type { FetchListOptions } from '../../../organizations/hooks'
+import { organizationMessages as messages } from '../../utils/messages'
 
 export const TagVariantByStatus: Record<string, TagProps['variant']> = {
   active: 'green',
@@ -12,9 +14,14 @@ export const TagVariantByStatus: Record<string, TagProps['variant']> = {
 
 interface Props {
   onSearch: (options: Partial<FetchListOptions>) => void
+  customFieldName?: string
 }
 
-const OrganizationsListSearch: React.FC<Props> = ({ onSearch }) => {
+const OrganizationsListSearch: React.FC<Props> = ({
+  onSearch,
+  customFieldName = '',
+}) => {
+  const { formatMessage } = useIntl()
   const search = useSearchState()
 
   const handleSearchKeyDown: React.KeyboardEventHandler<HTMLFormElement> = event => {
@@ -37,12 +44,19 @@ const OrganizationsListSearch: React.FC<Props> = ({ onSearch }) => {
     })
   }
 
+  const placeholder = customFieldName
+    ? formatMessage(messages.searchCustomFieldPlaceholder, {
+        fieldName: customFieldName,
+      })
+    : formatMessage(messages.searchPlaceholder)
+
   return (
     <Search
       rel=""
       {...inputProps}
       onClear={handleSearchClear}
       onKeyDown={handleSearchKeyDown}
+      placeholder={placeholder}
       className={csx({ marginRight: '$space-5' })}
     />
   )