+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Organizations admin: dropdown to search organizations by organization custom field value (B2BTEAM-3594).
+
 ## [3.2.2] - 2026-09-02
 
 ### Changed